+++ steinbock/_cli/apps.py
@@ -40,22 +40,6 @@ def ilastik_cmd(ilastik_binary, ilastik_args, ilastik_env):
 
 
 @apps_cmd_group.command(
-    name="cellpose",
-    context_settings={"ignore_unknown_options": True},
-    help="Run cellpose GUI",
-    add_help_option=False,
-)
-@click.argument("cellpose_args", nargs=-1, type=click.UNPROCESSED)
-@click_log.simple_verbosity_option(logger=steinbock_logger)
-@catch_exception(handle=SteinbockException)
-@use_ilastik_env
-def ilastik_cmd(cellpose_args, ilastik_env):
-    args = ["cellpose"] + list(cellpose_args)
-    result = run_captured(args, env=ilastik_env)
-    sys.exit(result.returncode)
-
-
-@apps_cmd_group.command(
     name="cellprofiler",
     context_settings={"ignore_unknown_options": True},
     help="Run CellProfiler GUI",
@@ -162,5 +146,49 @@ def jupyterlab_cmd(python_path, jupyter_module, jupyterlab_args):
     if not any(arg.startswith("--ip=") for arg in jupyterlab_args):
         jupyterlab_args.append("--ip='0.0.0.0'")
     args = [python_path, "-m", jupyter_module, "lab"] + jupyterlab_args
+    result = run_captured(args)
+    sys.exit(result.returncode)
+
+
+
+
+@apps_cmd_group.command(
+    name="cellpose",
+    context_settings={"ignore_unknown_options": True},
+    help="Run cellpose GUI",
+    add_help_option=False,
+)
+@click.option(
+    "--python",
+    "python_path",
+    type=click.Path(dir_okay=False),
+    default="/opt/cellprofiler-venv/bin/python",
+    show_default=True,
+    help="Python path",
+)
+@click.option(
+    "--cellprofiler",
+    "cellprofiler_module",
+    type=click.STRING,
+    default="cellprofiler",
+    show_default=True,
+    help="CellProfiler module",
+)
+@click.option(
+    "--plugins-directory",
+    "cellprofiler_plugin_dir",
+    type=click.Path(file_okay=False),
+    default="/opt/cellprofiler_plugins",
+    show_default=True,
+    help="Path to the CellProfiler plugin directory",
+)
+#@click.argument("cellprofiler_args", nargs=-1, type=click.UNPROCESSED)
+@click_log.simple_verbosity_option(logger=steinbock_logger)
+@catch_exception(handle=SteinbockException)
+def cellpose_view_cmd(
+    python_path, cellprofiler_module, cellprofiler_plugin_dir
+):
+    args = ["python", "-m", "cellpose"]
+
     result = run_captured(args)
     sys.exit(result.returncode)