+++ steinbock/segmentation/_cli/_cli.py
@@ -4,6 +4,7 @@ from ..._cli.utils import OrderedClickGroup
 from .cellpose import cellpose_cli_available, cellpose_cmd
 from .cellprofiler import cellprofiler_cmd_group
 from .deepcell import deepcell_cli_available, deepcell_cmd
+from .train import train_cmd_group
 
 
 @click.group(
@@ -14,7 +15,7 @@ from .deepcell import deepcell_cli_available, deepcell_cmd
 def segment_cmd_group():
     pass
 
-
+segment_cmd_group.add_command(train_cmd_group)
 segment_cmd_group.add_command(cellprofiler_cmd_group)
 if deepcell_cli_available:
     segment_cmd_group.add_command(deepcell_cmd)