+++ steinbock/segmentation/_cli/cellpose.py
@@ -5,7 +5,7 @@ import click_log
 import numpy as np
 
 from ... import io
-from ..._cli.utils import catch_exception, logger
+from ..._cli.utils import OrderedClickGroup, catch_exception, logger
 from ..._steinbock import SteinbockException
 from ..._steinbock import logger as steinbock_logger
 from .. import cellpose
@@ -14,6 +14,7 @@ cellpose_cli_available = cellpose.cellpose_available
 
 
 @click.command(name="cellpose", help="Run an object segmentation batch using Cellpose")
+
 @click.option(
     "--model",
     "model_name",
@@ -35,7 +36,7 @@ cellpose_cli_available = cellpose.cellpose_available
             "LC4",
         ]
     ),
-    default="cyto2",
+    default="tissuenet",
     show_default=True,
     help="Name of the Cellpose model",
 )
@@ -47,6 +48,7 @@ cellpose_cli_available = cellpose.cellpose_available
     show_default=True,
     help="Path to the image directory",
 )
+
 @click.option(
     "--minmax/--no-minmax",
     "channelwise_minmax",
@@ -197,8 +199,8 @@ def cellpose_cmd(
     img_files = io.list_image_files(img_dir)
     Path(mask_dir).mkdir(exist_ok=True)
     for img_file, mask, flow, style, diam in cellpose.try_segment_objects(
-        model_name,
         img_files,
+        model_name,
         channelwise_minmax=channelwise_minmax,
         channelwise_zscore=channelwise_zscore,
         channel_groups=channel_groups,
@@ -218,3 +220,10 @@ def cellpose_cmd(
         mask_file = io._as_path_with_suffix(Path(mask_dir) / img_file.name, ".tiff")
         io.write_mask(mask, mask_file)
         logger.info(mask_file)
+
+
+
+
+
+#@click_log.simple_verbosity_option(logger=steinbock_logger)
+#@catch_exception(handle=SteinbockException)