+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] - 2023-08-03
+
+### Added
+
+  - New `steinbock segment train prepare` command
+  - New `steinbock segment train prepare` command
+  - New `steinbock apps cellpose` command
+  - Added opencv-python==4.8.0.74 to requirement.txt
+
+### Changed
+
+  - The choice of pre-trained models in cellpose segmentation is expanded to include the pre-trained models introduced in cellpose 2
+  - Updated cellpose default parameters for better default segmentation results
+  - Updated documentation
+
 ## [0.16.0] - 2023-02-03
 
 ### Added